requirements.txt
Flask==2.1.0
google-auth==1.27.0
google-auth-oauthlib==0.4.2
oauthlib==3.1.0
pathlib==1.0.1
vendor==0.1
requests==2.25.1
requests-oauthlib==1.3.0
flask_sqlalchemy